More robust platform definition support.
The idea is that someday it will be possible to just concatenate the entire
set of source files into a single giant source file, for systems that
want to work this way.  As a result, the build system now compiles every
file, although some of them will not have any definitions.