Change in handling of extended info for messages.
Instead of supplying a pipe, and expecting that the info there would be included
we use nng_msg_getopt(). This will be enabled by the app asking for extended
information by setting an option, we don't copy the data for every app (most won't
care).  This means we don't have to worry about reference counting the pipe for
the life of associated messages.