Pipe ID calculations & socket error functions.
The use of platform_next_id was a bit off, because it could give
back pipe IDs that were too large (the high order bit must be
clear), and in very long running applications serving many
connections, the IDs could wrap and lead to duplicates.

Also we have added functions to set the recverr or senderr values,
which can be used by protocols -- either during initialization,
or during filters.  (REQ uses this for example.)