Implementation of an id hash for hashing pipes by ID.
We use some hints from Python's dict implementation, using an
open addressing scheme, and just ripping off the lower bits as
needed.  Since we assign IDs consecutively, this should work
well.  We shrink the table when it is only 1/8 full, and
we ensure that we grow the table when it is 2/3 full.  (The
growth will start by at minimum doubling the required size.)