Fix TCP hostname wildcards on Windows.  Sort of.
I seem to be having a very difficult time getting dual-stack sockets
to function properly on Windows.  I've sort of abandoned it for now.
I need to think about how to solve this -- it's not clear to me
right now whether dual stack sockets are the right answer or not.
People do expect these to work, but a tcp6:// url might be more
elegant.