Add endpoint tuning of maxrcv size.  Fix cmsg API.
The CMSG handling was completely borked.  This is fixed now, and
we stash the SP header size (ugh) in the CMSG contents to match what
nanomsg does.  We now pass the cmsg validation test.

We also fixed handling of certain endpoint-related options, so that
endpoints can get options from the socket at initialization time.
This required a minor change to the transport API for endpoints.

Finally, we fixed a critical fault in the REP handling of RAW sockets,
which caused them to always return NNG_ESTATE in all cases.  It should
now honor the actual socket option.