Add device support & testing.  Bus semantic fix.
This adds nn_device and nng_device.  There were some internal changes
required to fix shutdown / close issues.  Note that we shut down the
sockets when exiting from device -- this is required to make both threads
see the failure and bail, since we are not using a single event loop.

I also noticed that the bus protocol had a bug where it would send
messages back to the originator.  This was specifically tested for in
the compat_device test, and we have fixed it.