Convert to POSIX polled I/O for async; start of cancelable aio.
This eliminates the two threads per pipe that were being used to provide
basic I/O handling, replacing them with a single global thread for now,
that uses poll and nonblocking I/O.  This should lead to great scalability.

The infrastructure is in place to easily expand to multiple polling worker
threads.  Some thought needs to be given about how to scale this to engage
multiple CPUs.  Horizontal scaling may also shorten the poll() lists easing
C10K problem.

We should look into better solutions than poll() for platforms that have
them (epoll on Linux, kqueue on BSD, and event ports on illumos).

Note that the file descriptors start out in blocking mode for now, but
then are placed into non-blocking mode.  This is because the negotiation
phase is not yet callback driven, and so needs to be synchronous.