Fix close-related leak of pipes.
We have seen leaks of pipes causing test failures (e.g. the Windows
IPC test) due to EADDRINUSE.  This was caused by a case where we
failed to pass the pipe up because the AIO had already been canceled,
and we didn't realize that we had oprhaned the pipe.  The fix is to
add a return value to nni_aio_finish, and verify that we did finish
properly, or if we did not then we must free the pipe ourself.  (The
zero return from nni_aio_finish indicates that it accepts ownership
of resources passed via the aio.)