Always run the AIO completion logic.
We have seen some yet another weird situation where we had an orphaned
pipe, which was caused by not completing the callback.  If we are going
to run nni_aio_fini, we should still run the callback (albeit with a
return value of NNG_ECANCELED or somesuch) to be sure that we can't
orphan stuff.