Use a dedicated reap thread instead of taskq.
The problem is that reaping these things performs some blocking
operations which can tie up slots in the taskq, preventing other
tasks from running.  Ultimately this can lead to a deadlock as
tasks that are blocked wind up waiting for tasks that can't get
scheduled.  Blocking tasks really should not run on the system taskq.