Fix crash when using legacy send API with NN_MSG.
We introduced the compat_msg.c from the old msg.c in the nanomsg
repo.  While here, we found that the handling for send() was badly
wrong, by a level of indirection.  We simplified the code to so that
nn_send() and nn_recv() are simple wrappers around the nn_sendmsg()
and nn_recvmsg() APIs (as in old nanomsg).  This may not be quite as
fast, but it's more likely to be correct and reduces complexity.