Add some more compatibility tests; fix surveyor compat bug.
We noticed a bug in the surveyor handling of the options; this fixes
that.  At the same time, we noticed a race condition in the setting
of the error for future calls, a short sleep seems to cure it.  This
distinction (ESTATE vs ETIMEDOUT) is pretty annoying, and it would be
better to have a different way to handle it.  More work here is
warranted.