Updates to Travis CI configuration, serialize testing (TCP).
For now we have problems with TCP tests racing against each
other and this can lead to failures due to port use.  In the
old nanomsg repo we assigned several different port numbers on
the command line, but we lack that here generally.  However, we
feel that parallel testing is of little merit; its better to
build dedicated stress tests if that is what is desired.

We also use the matrix stuff a bit more sensibly, only installing
packages we actually use. This will likely lead to shorter startup
times.

Finally we cleaned up the environment so that we have sane defaults
for most environment variables.