Add new PAIR_V1 protocol.
The PAIR_V1 protocol supports both raw and cooked modes, and has loop
prevention included.  It also has a polyamorous mode, wherein it allows
multiple connections to be established.  In polyamorous mode (set by
an option), the sender requests a paritcular pipe by setting it on the
message.

We default to PAIR_V1 now.