Unify the msg API.
This makes the operations that work on headers start with
nni_msg_header or nng_msg_header.  It also renames _trunc to
_chop (same strlen as _trim), and renames prepend to insert.

We add a shorthand for clearing message content, and make
better use of the endian safe 32-bit accessors too.

This also fixes a bug in inserting large headers into messages.

A test suite for message handling is included.