Convenience option accesor functions.
This adds functions that know about option sizes and make them
easier to use.  While here I added some validation of those, and
cleaned up a few tests slightly.  Note that we do not need to
use the nng_impl.h for most tests.  More of them need to be
cleaned up.