fixes #62 Endpoint close should be synchronous #62 fixes #66 Make pipe and endpoint structures private
This changes a number of things, refactoring endpoints and supporting
code to keep their internals private, and making endpoint close
synchronous.  This will allow us to add a consumer facing API for
nng_ep_close(), as well as property APIs, etc.

While here a bunch of convoluted and dead code was cleaned up.