Add support for 64-bit ids in idhash.
We intend to use this with transports where dynamic "port numbers"
might be 32-bits.  This would allow us to formulate a 64-bit number
representing a conversation, and be able to find that conversation
by the 64-bit value.

Note that the hashed values are probably not perfectly optimal, as
only the low order bits are particularly significant in the hash.
We might want to consider XOR'ing in the upper bits to address that.