Eliminate legacy option settings, provide easier option IDs.
This eliminates all the old #define's or enum values, making all
option IDs now totally dynamic, and providing well-known string
values for well-behaved applications.

We have added tests of some of these options, including lookups, and
so forth.  We have also fixed a few problems; including at least
one crasher bug when the timeouts on reconnect were zero.

Protocol specific options are now handled in the protocol.  We will
be moving the initialization for a few of those well known entities
to the protocol startup code, following the PAIRv1 pattern, later.

Applications must therefore not depend on the value of the integer IDs,
at least until the application has opened a socket of the appropriate
type.