More pipe option handling, pipe API support.  Url option.
This fleshes most of the pipe API out, making it available to end user
code.  It also adds a URL option that is independent of the address
options (which would be sockaddrs.)

Also, we are now setting the pipe for req/rep.  The other protocols need
to have the same logic added to set the receive pipe on the message.  (Pair
is already done.)