Refactor option handling APIs.
This makes the APIs use string keys, and largely eliminates the use of
integer option IDs altogether.  The underlying registration for options
is also now a bit richer, letting protcols and transports declare the
actual options they use, rather than calling down into each entry point
carte blanche and relying on ENOTSUP.

This code may not be as fast as the integers was, but it is more intuitive,
easier to extend, and is not on any hot code paths.  (If you're diddling
options on a hot code path you're doing something wrong.)