Added more complete tests, and changes to property handling.
We allow some properties to be set on endpoints after they are
started; transports now responsible for checking that.  (The new
values will only apply to new connections of course!)

We added short-hand functions for pipe properties, and also added
uint64_t shorthands across the board.

The zerotier documentation got some updates (corrections).  We have
also added a separate header now for the ZT stuff.

Also, dialers and listeners do not intermix anymore -- we test that
only a dialer can be used with setting dialer options, and likewise
for listeners.