New platform API for storage methods.
This includes converting the ZeroTier transport to use these.

The new API supports file creation, retrieval, and deletion. It
also supports directory methods for traversal, creation, and
deletion.  It also has a few methods to obtain well-known directories
like $TMPDIR and $HOME.

A rich test suite for this functionality is added as well.