fixes #45 expose aio to applications
While here we added a test for the aio stuff, and cleaned up some dead
code for the old fd notifications.  There were a few improvements to
shorten & clean code elsewhere, such as short-circuiting task wait
when the task has no callback.

The legacy sendmsg() and recvmsg() APIs are still in the socket core
until we convert the device code to use the aios.