fixes #46 make device() use aios directly
This eliminates the separate threads used for devices, letting them
benefit from the new aio framework.  It also eliminates the legacy
nni_sock_sendmsg and nni_sock_recvmsg internal APIs.

It would appear that there is an opportunity here to provide asynchronous
device support out to userland as well, exposing an aio to them.  That
work is deferred to later.