Documentation updates.
This adds a few man pages, fixes the content and formatting on a few
others.

It also adds two utility scripts for me to use, publish.sh, which I
use to publish updates to the gh-pages branch, and preview.sh, which
I use to look at markup without cluttering my work directory.