CI updates.
This updates to use Circle CI by default, but we leave Travis in place
(but only one default build and one coverage build) until we can move it.
We also switch to using ninja for performance reasons.

The hope is to get to using CircleCI for everything except Windows.  The
AppVeyor build needs cleanups and should switch to ninja later.