Fix websocket hang after sending one message.
This fixes a problem where the websocket would only send one message,
then no others, due to not clearing the "frame" busy flag on completion
of the frame transmit.

We have also added a test that tries to send 10 messages back and
forth to make sure that we catch this kind of problem in the future.

Finally we've fixed some problems that were found when testing edge
cases around the protocol, which were responsible for invalid memory
accesses.