Refactored file API.
This refactor of the file API provides a simpler and easier to use
interface for our needs (and simpler to implement) in both the
ZeroTier transport and the HTTP/TLS file accesses.  It also removes
some restrictions present on the old one, although it is still not
suitable for working with large files.  (It will work, just be
very inefficient as the entire file must be loaded into memory.)