fixes #216 HTTP server side API refactoring, directory serving support
This changes the backend (internal) HTTP API to provide a much more
sensible handler scheme, where the handlers are opaque objects and we
can allocate a handler for different types of tasks.

We've also added support serving up directories of static content, and
added code to validate that the directory serving is working as intended.

This is a key enabling step towards the public API.