Add nng_opts_parse() API for handling command line options.
We have implemented this alternative to getopt() so that we can
create nngcat.  The reason we did not just use getopt() is that
getopt() does not understand long options (which nanocat uses, and
we want to preserve for compatibility) and getopt() is not available
on Windows (and possibly other non-POSIX platforms.)

This function handles long and short options, but does not have support
for option clustering.  It also is threadsafe & reentrant, unlike getopt.