Remove some attribute declarations, use build scripts.
This is phase 1.  We still have the copyright text in place for now,
but I intend to replace that too.  The reason for this is to facilitate
efforts to build a larger reference manual as a single PDF book file.

The other problem I have to solve is the fact that man page chapters wind
up creating page breaks, which is not desirable for a book form.