fixes #240 nngcat is MIA
This is intended to provide compatibility with, and has been tested
against, legacy nanocat.  There are a few differences though.

At this time support for the alias names (where argv[0] is set to
something like nngreq or somesuch) is missing.

By default this library operations without NNG_FLAG_NONBLOCK on
dial and listen, so that failures here are immediately diagnosable.
(This behavior can be changed with the --async flag.)

By default --pair means PAIRv1, but you can specify --pair0
or --pair1 explicitly.  (There is also a --compat mode, and in
that mode --pair means PAIRv0.  The --compat mode also turns on
NNG_FLAG_NONBLOCK by default.)

The "quoted" mode also quotes tabs.  (Legacy nanocat did not.)

It is possible to connect to *multiple* peers by using the --dial
or --listen (or similar) options multiple times.

Shorthands can be used for long options that are not ambiguous.  For
example, --surv can be used to mean surveyor, but --re is invalid because
it can mean req, rep, or respondent.

We assume you have a reasonable standard C environment.  This won't work
in embedded environments without support for FILE *.

TLS options are missing but to be added soon.

A man page is still to be written.