fixes #269 nngcat unreliable with ZeroTier transport
This does a few things.  First it closes a case where a dropped
message could prevent subsequent connection attempts from getting through.

Second, it changes the rate at which we retry, and the timeout, to be
a lot more aggressive when attempting to establish a connection.  We
retry every 500 ms, for up to 2 minutes, before giving up.  This gives
a lot more resilience in the face of message loss that is typical of
ZeroTier in some environments when first establishing communication.

Third, makes the values for the connection attempts *tunable*, so
that applications can adjust for different deployment scenarios.

Fourth, it includes the ability to get the UDP socket name.  This was
needed during some debug, and may be useful for a real UDP transport
later, so we're keeping it.

Finally, we added documentation for the above items.