fixes #296 Typed options should validate option type fixes #302 nng_dialer/listener/pipe_getopt_sockaddr desired
This adds plumbing to pass and check the type of options
all the way through.

NNG_ZT_OPT_ORBIT is type UINT64, but you can use the untyped form to
pass two of them if needed.

No typed access for retrieving strings yet.  I think this should allocate
a pointer and copy that out, but that's for later.