fixes #317 TLS, TCP, and IPC should support multiple outstanding ops fixes #22 Consider using synchronous completions sometimes
Transport improvements for IPC, TCP, and TLS.

This change does three things.

First it permits multiple outstanding receives or sends on the transport.
This change is being made to accomodate some other changes in the protocols
where it might be advantageous to post send or receives directly against
the transport pipe without going through another level of indirection.

Second, it changes the normal completions to be performed synchronously.
This translates into a rather major performance improvement, reducing
latency by some 27%, and thereby improving performance altogether.  (This
elminates two extra context switches per transaction!)

FInally, we can save some extra checks and conditions because we know
that completions cannot happen if we don't have a pending operation
(we no longer complete out of sequence), and we only call the dosend
operation when we have something to send.  This can eliminate some
pipeline stalls.