fixes #331 replace NNG_OPT_RAW option with constructor
This makes the raw mode something that is immutable, determined
at socket construction.  This is an enabling change for the
separate context support coming soon.

As a result, this is an API breaking change for users of the raw
mode option (NNG_OPT_RAW).  There aren't many of them out there.

Cooked mode is entirely unaffected.

There are changes to tests and documentation included.