fixes #355 Possible use after free in REP
While here I've added some code that should help us backtrack on
a crash, by linking back to the pipe from the context when we
are queued on that pipes sendq.

I'm not sure if we've ever seen these or not, but it could explain
certain infrequent crashes we think we've seen.