Markup fixes, ensuring links are colored properly.
It turns out that when creating cross references, we need to
make any text styling (generally literal characters) outside of
the link, to avoid having the styling override the link color.
(We prefer to have links colored for ease of use.)

While here a few other markup, and actual content, errors were fixed.