Fix failures in pipe test on Windows.
The problem was that with inproc we could wind up calling the
completion callback before actually returning the dialer ID
to the structure, leading to false comparison failures.

Creating the dialer and listener in separate steps ensures that
we have IDs in place for them before any callback might be issued.
Note that this was a bug in the test logic only, and not a problem
with the underlying implementation.